Converted Token Generator and Updated Login (#11)
